modified:   include/SerialMultiplexer.h 	modified:   src/CarController.cpp 	modified:   src/main.cpp

+++ include/SerialMultiplexer.h
@@ -1,13 +1,15 @@
 #ifndef SERIAL_MULTIPLEXER_H
 #define SERIAL_MULTIPLEXER_H
 
+#include <stdint.h>
+
 class SerialMultiplexer
 {
 public:
     void tcaSelect(uint8_t i); // Function to select TCA9548A channel
 
 private:
-    const int TCAADDR = 0x70; // I2C address of the TCA9548A
+    const uint8_t TCAADDR = 0x70; // I2C address of the TCA9548A
 };
 
-#endif
+#endif // SERIAL_MULTIPLEXER_H
